Load the bundle
